Replaced getLatestPosts usage (wp.api.collections.Posts wrapper) with withAPIData HoC in latestPosts.
